requirements.txt
requests>=2.31.0
PyYAML>=6.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
